preliminary support for shader sets in Dolphin
Primarily a placeholder, but if it's possible to patch in extra shaders at no maintainence cost then this list could be expanded to include mad-scanlines-crt for the scanline sets too.

It doesn't seem feasible to support actual sets in Dolphin as Dolphin doesn't seem to be able to load them.